feat: 用户退出登录功能开发

+++ user-core/src/main/java/top/easyblog/titan/service/impl/LoginServiceImpl.java
@@ -4,6 +4,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
@@ -11,10 +12,13 @@ import javax.servlet.http.HttpServletRequest;
 
 import top.easyblog.titan.annotation.Transaction;
 import top.easyblog.titan.bean.LoginDetailsBean;
+import top.easyblog.titan.bean.SignInLogBean;
 import top.easyblog.titan.bean.UserDetailsBean;
 import top.easyblog.titan.constant.LoginConstants;
+import top.easyblog.titan.enums.Status;
 import top.easyblog.titan.exception.BusinessException;
 import top.easyblog.titan.request.LoginRequest;
+import top.easyblog.titan.request.QuerySignInLogRequest;
 import top.easyblog.titan.request.RegisterUserRequest;
 import top.easyblog.titan.response.ResultCode;
 import top.easyblog.titan.service.ILoginService;
@@ -34,8 +38,13 @@ public class LoginServiceImpl implements ILoginService {
 
     @Autowired
     private AccessAccountService accessAccountService;
+
     @Autowired
     private AccessPhoneAuthService phoneAuthService;
+
+    @Autowired
+    private UserSignInLogService signInLogService;
+
     @Autowired
     private RedisService redisService;
 
@@ -69,15 +78,26 @@ public class LoginServiceImpl implements ILoginService {
     }
 
     @Override
-    public void logout(String token) {
+    public void logout(String token, HttpServletRequest httpServletRequest) {
         if (StringUtils.isEmpty(token)) {
             throw new BusinessException(ResultCode.REQUIRED_PARAM_TOKEN_NOT_EXISTS);
         }
+        String userInfoJsonStr = redisService.get(token);
+        if (StringUtils.isEmpty(userInfoJsonStr)) {
+            //检查是否过期或者token无效，如果token找不到直接返回
+            return;
+        }
+        UserDetailsBean userDetailsBean = JsonUtils.parseObject(userInfoJsonStr, UserDetailsBean.class);
+        Long id = userDetailsBean.getId();
+        
+        List<SignInLogBean> signInLogs = userDetailsBean.getSignInLogs();
+        SignInLogBean signInLogBean = signInLogService.querySignInLogDetails(QuerySignInLogRequest.builder()
+                .userId(id).status(Status.ENABLE.getCode()).build());
         redisService.delete(token);
     }
 
     @Override
-    public UserDetailsBean register(RegisterUserRequest request) {
+    public UserDetailsBean register(RegisterUserRequest request, HttpServletRequest httpServletRequest) {
         LoginPolicy loginPolicy = LoginPolicyFactory.getLoginPolicy(request.getIdentifierType());
         if (Objects.isNull(loginPolicy)) {
             throw new BusinessException(ResultCode.INTERNAL_ERROR);
